Small code cleanup
git-svn-id: svn://scribus.net/branches/Version15x/Scribus@25661 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/canvas.cpp
@@ -1471,7 +1471,7 @@ void Canvas::DrawPageItems(ScPainter *painter, ScLayer& layer, QRect clip, bool 
 
 	PageItem *currItem;
 	int layerCount = m_doc->layerCount();
-	int docCurrPageNo=static_cast<int>(m_doc->currentPageNumber());
+	int docCurrPageNo = m_doc->currentPageNumber();
 	if ((layerCount > 1) && ((layer.blendMode != 0) || (layer.transparency != 1.0)) && (!layer.outlineMode))
 		painter->beginLayer(layer.transparency, layer.blendMode);
